pyproject.toml
[tool.poetry]
name = "schola"
version = "0.1.0"
description = "Student-Centric Hub Online Learning Assistant"
authors = ["Reagan <rrr.chanhiulok@gmail.com>"]
license = "Proprietary"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
openai = "^1.47.1"
requests = "^2.32.3"
python-dotenv = "^1.0.1"
python-telegram-bot = "^21.6"
azure-ai-formrecognizer = "^3.3.3"
pydantic = "^2.9.2"
replicate = "^0.34.1"
faiss-cpu = "1.7.4"
pymupdf = "^1.24.10"
python-docx = "^1.1.2"
python-pptx = "^1.0.2"
pypdf = "^5.0.1"
pandas = "^2.2.3"
numpy = "1.26.4"
fastapi = "^0.115.0"
uvicorn = "^0.31.0"
python-multipart = "^0.0.12"


[tool.poetry.group.dev.dependencies]
ruff = "^0.6.7"
mypy = "^1.11.2"
pytest = "^8.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
